+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/mock/FakeGoogleStorageInterpreter.scala
@@ -87,9 +87,17 @@ class BaseFakeGoogleStorage extends GoogleStorageService[IO] {
                                 traceId: Option[TraceId] = None,
                                 retryConfig: RetryConfig,
                                 expirationTime: Long,
-                                expirationTimeUnit: TimeUnit
+                                expirationTimeUnit: TimeUnit,
+                                queryParams: Map[String, String]
   ): Stream[IO, URL] =
-    localStorage.getSignedBlobUrl(bucketName, blobName, signingCredentials, traceId)
+    localStorage.getSignedBlobUrl(bucketName,
+                                  blobName,
+                                  signingCredentials,
+                                  traceId,
+                                  expirationTime = expirationTime,
+                                  expirationTimeUnit = expirationTimeUnit,
+                                  queryParams = queryParams
+    )
 
   override def downloadObject(blobId: BlobId,
                               path: Path,